pyproject.toml
[project]
name = "slc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-allauth[socialaccount,mfa]==65.13.1",
    "django-ninja>=1.4.3",
    "django-storages[s3]>=1.14.6",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg[binary]==3.3.2",
    "python-slugify>=8.0.4",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "whitenoise>=6.11.0",
    "django-anymail[brevo]==14.0",
    "ruff>=0.14.1",
    "ffmpeg>=1.4",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
